Allow for custom_error_response argument.
This PR allows for an array of `custom_error_response` to be set when using the module.

The argument is referenced in the cloudfront_distribution docs at https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudfront_distribution

I am by no means an expert at tf, please see this PR as a suggestion for your consideration. It may well need iteration, or rewriting as you see fit. It's untested, though I do have a dev namespace where I can run a test provisioning.

The justification for this PR is that my team is serving a static site via CloudFront, currently, when a user needs to login they will a non-user-friendly error message.

After this PR we will be able to set a custom 403 error message telling the uset that they will need to log in to see the content.